+++ R/inRt.R
@@ -18,17 +18,17 @@
 #' and from 20 to 6 of the next day. 'civil.twilight' and 'activity.times' must be
 #' FALSE to allow the use of 'limits'.
 #' @param coor coordinates for the location where temperature was recorded,
-#' formatted as decimal degrees N/S, decimal degress E/W.
+#' formatted as decimal degrees N/S, decimal degrees E/W.
 #' When 'civil.twilight' is TRUE, 'coor' allows the user to define sunrise and sunset times
 #' based on the \code{\link{crepuscule}} function (in \code{maptools} package). 
 #' @param civil.twilight TRUE or FALSE. Set as TRUE when time periods for calculation
 #' are to be defined by civil twilight times - calculated using \code{\link{crepuscule}}. 
 #' If 'civil.twilight = TRUE', 'coor' and 'time.zone' need to be specified.
 #' @param activity.times TRUE or FALSE. Set as TRUE when time periods for calculation
-#' are defined by \code{\link{incRactivity}}. Data must contain a column named 
-#' 'incR_score' for the use of \code{\link{incRactivity}}.
+#' are defined by \code{\link{incRact}}. Data must contain a column named 
+#' 'incR_score' for the use of \code{\link{incRact}}.
 #' @param time.zone time zone for \code{\link{crepuscule}} dawn and dusk calculations.
-#' @param ... use parameters in \code{\link{incRactivity}} if \emph{activity.times} = TRUE.
+#' @param ... use parameters in \code{\link{incRact}} if \emph{activity.times} = TRUE.
 #' @return a data frame containing temperature means and variance for the defined time 
 #' window.
 #' @author Pablo Capilla-Lasheras
@@ -53,8 +53,8 @@
 #'         civil.twilight=FALSE, 
 #'         activity.times=TRUE,
 #'         time.zone=NULL,
-#'         time_column="time",             # extra argument needed for incRactivity
-#'         vector.incubation="incR_score") # extra argument needed for incRactivity
+#'         time_column="time",             # extra argument needed for incRact
+#'         vector.incubation="incR_score") # extra argument needed for incRact
 #'         
 #' # calculation based on civil twilight
 #' incRt (data=incR_procdata, 
@@ -64,7 +64,7 @@
 #'         civil.twilight=TRUE, 
 #'         activity.times=FALSE,
 #'         time.zone="GMT")
-#' @seealso \code{\link{incRprep}} \code{\link{incRscan}} \code{\link{incRactivity}}
+#' @seealso \code{\link{incRprep}} \code{\link{incRscan}} \code{\link{incRact}}
 #' \code{\link{crepuscule}}
 #' @export 
 
@@ -96,14 +96,14 @@ incRt <- function (data,
   # two periods of time within 24hrs.
   # (1) you specify the time window you want
   ## to compute day and night mean and variation; 
-  # (2) it takes first_offbout and last_onbout activity using incRactivity; or,
+  # (2) it takes first_offbout and last_onbout activity using incRact; or,
   # (3) uses civil twilight times to define night 
   # times to define day and night periods and calculate variation and temperature.
   #
   # First I create a table which specifies such periods depending on 1, 2 or 3.
   if (activity.times==TRUE) {
     # calculates first_offbout and last_onbout activity times
-    act.times <- incRactivity (data= data.onoff.act, 
+    act.times <- incRact (data= data.onoff.act, 
                                ...)
     act.times$first_offbout <- do.call(args = base::lapply(strsplit(act.times$first_offbout, " "), 
                                                            FUN = function(x) {
@@ -192,7 +192,7 @@ incRt <- function (data,
     # DAY CALCULATIONS
     #
     if (is.null(df00$dec_time)) {
-      stop("Dec_time column is missing. Please, use exactly that name.")
+      stop("'dec_time' column is missing. Please, use exactly that name.")
       }
     day.data <- df00[df00$dec_time > day.morning & df00$dec_time < day.evening, ]
     data.final$day.mean[k] <- base::mean (day.data[[temp.name]], na.rm=TRUE)