+++ NEWS.md
@@ -1,4 +1,16 @@
+# incR 1.0.1
+* bug in incRt fixed
+* version release to CRAN
+
+# incR 1.0.0.9000
+* incRconstancy changed by incRatt
+* incRactivity changed by incRact
+
+# incR 0.3.1.9000
+* Version for development available on GitHub
+
 # incR 0.3.1
+* 20/07/2016 v0.3.1 available on CRAN
 
 * Most of incRbouts has been coded again using a different approach to increase speed
 * New functionality included in incRbouts: per day and per bout outputs